requirements.txt
progress
numpy
pandas
seaborn
plotly
dash
pyserial
pymodbus
requests
pymongo
logging